Hoangtl

+++ Homework05/Projects/EBRApp/src/main/java/com/hust/ebr/beans/Bike.java
@@ -25,7 +25,7 @@ public class Bike {
     @JsonFormat(pattern="dd/MM/yyyy")
     private Date manufacturingDate;
     private String producer;
-    private float cost;
+    private long cost;
     private Status status;
     private String dockingStationId;
